+++ rssgenerator/LocalStore.py
@@ -9,6 +9,7 @@ _generator_name = __name__ + "-" + ".".join(map(str, __version__))
 
 from django.conf import settings
 from rssgenerator.models import Links
+from rssgenerator.tools import images
 
 import os
 import struct
@@ -80,95 +81,14 @@ class LocalStore:
         dstFile.write(response)
         dstFile.close()
 
-        infos = self.getHeightWidth(itemId, link)
+        infos = self.setHeightWidth(itemId, link)
         Links.objects.filter(id = link.id) \
                      .update(height = infos['h'],
                              width = infos['w'])
 
-
-    def getHeightWidth(self, itemId, link):
+    def setHeightWidth(self, itemId, link):
         if not link.storeLocaly:
             return {}
              
         linkFilePath = self.__getLinkFilePath(itemId, link)
-        if not os.path.exists(linkFilePath):
-            return {}
-        
-        infos = {}
-        (infos['w'], infos['h']) = get_image_size(linkFilePath)
-        return infos
-        
-
-#-------------------------------------------------------------------------------
-# Name:        get_image_size
-# Purpose:     extract image dimensions given a file path using just
-#              core modules
-#
-# Author:      Paulo Scardine (based on code from Emmanuel VAÏSSE)
-#
-# Created:     26/09/2013
-# Copyright:   (c) Paulo Scardine 2013
-# Licence:     MIT
-#-------------------------------------------------------------------------------
-class UnknownImageFormat(Exception):
-    pass
-
-def get_image_size(file_path):
-    """
-    Return (width, height) for a given img file content - no external
-    dependencies except the os and struct modules from core
-    """
-    size = os.path.getsize(file_path)
-
-    with open(file_path) as input:
-        height = -1
-        width = -1
-        data = input.read(25)
-
-        if (size >= 10) and data[:6] in ('GIF87a', 'GIF89a'):
-            # GIFs
-            w, h = struct.unpack("<HH", data[6:10])
-            width = int(w)
-            height = int(h)
-        elif ((size >= 24) and data.startswith('\211PNG\r\n\032\n')
-              and (data[12:16] == 'IHDR')):
-            # PNGs
-            w, h = struct.unpack(">LL", data[16:24])
-            width = int(w)
-            height = int(h)
-        elif (size >= 16) and data.startswith('\211PNG\r\n\032\n'):
-            # older PNGs?
-            w, h = struct.unpack(">LL", data[8:16])
-            width = int(w)
-            height = int(h)
-        elif (size >= 2) and data.startswith('\377\330'):
-            # JPEG
-            msg = " raised while trying to decode as JPEG."
-            input.seek(0)
-            input.read(2)
-            b = input.read(1)
-            try:
-                while (b and ord(b) != 0xDA):
-                    while (ord(b) != 0xFF): b = input.read(1)
-                    while (ord(b) == 0xFF): b = input.read(1)
-                    if (ord(b) >= 0xC0 and ord(b) <= 0xC3):
-                        input.read(3)
-                        h, w = struct.unpack(">HH", input.read(4))
-                        break
-                    else:
-                        input.read(int(struct.unpack(">H", input.read(2))[0])-2)
-                    b = input.read(1)
-                width = int(w)
-                height = int(h)
-            except struct.error:
-                raise UnknownImageFormat("StructError" + msg)
-            except ValueError:
-                raise UnknownImageFormat("ValueError" + msg)
-            except Exception as e:
-                raise UnknownImageFormat(e.__class__.__name__ + msg)
-        else:
-            raise UnknownImageFormat(
-                "Sorry, don't know how to get information from this file."
-            )
-
-    return width, height
+        return images.getHeightWidth(linkFilePath)
